added owner

+++ controllers/contacts.js
@@ -18,9 +18,14 @@ const listContacts = async (req, res) => {
 };
 
 const getContactById = async (req, res) => {
+  const { _id: owner } = req.user;
+  console.log(owner);
   const { contactId } = req.params;
-  const result = await Contact.findById(contactId);
-  if (!result) {
+
+  const result = await Contact.findById({ _id: contactId, owner });
+  console.log(result.owner.equals(owner));
+
+  if (!result || !result.owner.equals(owner)) {
     throw HttpError(404, "Not found");
   }
   res.status(200).json(result);
@@ -33,20 +38,27 @@ const addContact = async (req, res) => {
 };
 
 const removeContact = async (req, res) => {
+  const { _id: owner } = req.user;
   const { contactId } = req.params;
-  const result = await Contact.findByIdAndRemove(contactId);
-  if (!result) {
+  const result = await Contact.findByIdAndRemove({ _id: contactId, owner });
+  console.log(result);
+  if (!result || !result.owner.equals(owner)) {
     throw HttpError(404, "Not found");
   }
   res.status(200).json({ message: "contact deleted" });
 };
 
 const updateContact = async (req, res) => {
+  const { _id: owner } = req.user;
   const { contactId } = req.params;
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
-    new: true,
-  });
-  if (!result) {
+  const result = await Contact.findByIdAndUpdate(
+    { _id: contactId, owner },
+    req.body,
+    {
+      new: true,
+    }
+  );
+  if (!result || !result.owner.equals(owner)) {
     throw HttpError(404, "Not found");
   }
 
@@ -54,11 +66,17 @@ const updateContact = async (req, res) => {
 };
 
 const updateStatusContact = async (req, res) => {
+  const { _id: owner } = req.user;
   const { contactId } = req.params;
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
-    new: true,
-  });
-  if (!result) {
+  const result = await Contact.findByIdAndUpdate(
+    { _id: contactId, owner },
+    req.body,
+    {
+      new: true,
+    }
+  );
+
+  if (!result || !result.owner.equals(owner)) {
     throw HttpError(404, "Not found");
   }
